Create create_files.sh
This script employs Atomsk to generate disordered and deformed structures in POSCAR format from the specified initial structure. It selects the degree of atomic position disturbance and the extent of deformation based on a uniform distribution within a given range. Upon determining the atomic disturbance value, Atomsk applies a normal distribution to execute the modifications.