Create create_poscars.sh
For each file generated by the 'create_files.sh', this script extracts the filename and creates a corresponding directory within the target directory. Then, it copies the original file to the newly created directory, renaming it to 'POSCAR', and also copies additional required VASP files (INCAR, KPOINTS) into the same directory. POTCAR can be copied by the next illustrated script (which automatically prepares it with the potentials for found elements within POSCAR), or can be uncommented in this script and copied directly also.