Create create_files_2.sh
The second version of this script will also attempt to distribute the atoms such that no two are closer to each other than the specified 'distance_separ_limit' and it requires Python with Pandas module. Additionally, it generates an output file named 'output_all.txt', where the first column contains the names of the generated structures and the second column lists the minimum distances between atoms, as determined by the RDF for each structure.